Feat: Add DTOs for order and cart components 	*Refactor prices to BigDecimal

+++ aurora-backend/src/main/java/com/github/gr/aurora_bookstore/model/entity/Order.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -23,7 +25,7 @@ public class Order {
     private User user;
 
     @Column(name = "total_price", nullable = false)
-    private Float totalPrice;
+    private BigDecimal totalPrice;
 
     @Column(name = "date", insertable = false, updatable = false)
     private java.time.LocalDateTime date;
